+++ routes/api.php
@@ -1,7 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use PictaStudio\Contento\Http\Controllers\{ContentTagController, FaqCategoryController, FaqController, MailFormController, MenuController, MenuItemController, ModalController, PageController, SettingController};
+use PictaStudio\Contento\Http\Controllers\{ContentTagController, FaqCategoryController, FaqController, MailFormController, MenuController, MenuItemController, MetadataController, ModalController, PageController, SettingController};
 
 Route::apiResource('pages', PageController::class);
 Route::apiResource('menus', MenuController::class);
@@ -13,6 +13,9 @@ Route::post('faqs/bulk/upsert', [FaqController::class, 'upsertMultiple']);
 Route::apiResource('mail-forms', MailFormController::class);
 Route::apiResource('modals', ModalController::class);
 Route::apiResource('content-tags', ContentTagController::class);
+Route::apiResource('metadata', MetadataController::class)->parameters([
+    'metadata' => 'metadata',
+]);
 
 Route::get('settings', [SettingController::class, 'index']);
 Route::post('settings', [SettingController::class, 'store']);