+++ src/Http/Resources/MetadataResource.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace PictaStudio\Contento\Http\Resources;
+
+use Illuminate\Http\Request;
+
+class MetadataResource extends ContentoJsonResource
+{
+    public function toArray(Request $request): array
+    {
+        return [
+            'id' => $this->id,
+            'name' => $this->name,
+            'slug' => $this->slug,
+            'uri' => $this->uri,
+            'metadata' => $this->metadata,
+            'created_at' => $this->created_at,
+            'updated_at' => $this->updated_at,
+        ];
+    }
+}