+++ src/Validations/Contracts/MetadataValidationRules.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace PictaStudio\Contento\Validations\Contracts;
+
+interface MetadataValidationRules extends ProvidesValidationRules {}