+++ tests/TestCase.php
@@ -76,6 +76,7 @@ class TestCase extends Orchestra
             'create_content_tags_table.php',
             'update_content_tags_add_images.php',
             'create_content_taggables_table.php',
+            'create_metadata_table.php',
             'create_settings_table.php',
         ];